WIP Start setting up universal react app
This warrants a lot more explanation than I’m going to be able to put here.
My hope is that the code will speak for itself, there are a lot of examples of a universal app out there and I want this to be the cleanest one. 

This will take a lot of time.