Change api to user hapi.js
Converts api to use hapi.js instead of Express.
I've changed how `start-server.js` works, it now calls a function and gets the http
listener from that. This allows it to be a little more generic and it
works better with express. Also remove error handling route as hapi has
that all built in.

This fixes #4